// apps/earn-protocol/graphql/clients/top-depositors/client.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigDecimal: { input: any; output: any; }
  BigInt: { input: any; output: any; }
  Bytes: { input: any; output: any; }
  Int8: { input: any; output: any; }
  Timestamp: { input: any; output: any; }
};

export type Account = {
  __typename?: 'Account';
  claimedSummerToken: Scalars['BigInt']['output'];
  claimedSummerTokenNormalized: Scalars['BigDecimal']['output'];
  /**  Address of the account  */
  id: Scalars['ID']['output'];
  positions: Array<Position>;
};
// ...
export type Deposit = Event & {
  __typename?: 'Deposit';
  /**  Amount of token deposited in native units  */
  amount: Scalars['BigInt']['output'];
  /**  Amount of token deposited in USD  */
  amountUSD: Scalars['BigDecimal']['output'];
  /**  Token deposited  */
  asset: Token;
  /**  Block number of this event  */
  blockNumber: Scalars['BigInt']['output'];
  /**  Address that deposited tokens  */
  from: Scalars['String']['output'];
  /**  Transaction hash of the transaction that emitted this event  */
  hash: Scalars['String']['output'];
  /**  { Transaction hash }-{ Log index }  */
  id: Scalars['ID']['output'];
  /**  Amount of input token in the position  */
  inputTokenBalance: Scalars['BigInt']['output'];
  /**  Amount of input token in the position in USD  */
  inputTokenBalanceNormalizedUSD: Scalars['BigDecimal']['output'];
  /**  Event log index. For transactions that don't emit event, create arbitrary index starting from 0  */
  logIndex: Scalars['Int']['output'];
  /**  Position that this deposit belongs to  */
  position: Position;
  /**  The protocol this transaction belongs to  */
  protocol: YieldAggregator;
  /**  Timestamp of this event  */
  timestamp: Scalars['BigInt']['output'];
  /**  Address that received tokens  */
  to: Scalars['String']['output'];
  /**  The vault involving this transaction  */
  vault: Vault;
};
// ...
export type Event = {
  /**  Block number of this event  */
  blockNumber: Scalars['BigInt']['output'];
  /**  Address that sent tokens  */
  from: Scalars['String']['output'];
  /**  Transaction hash of the transaction that emitted this event  */
  hash: Scalars['String']['output'];
  /**  { Transaction hash }-{ Log index }  */
  id: Scalars['ID']['output'];
  /**  Event log index. For transactions that don't emit event, create arbitrary index starting from 0  */
  logIndex: Scalars['Int']['output'];
  /**  The protocol this transaction belongs to  */
  protocol: YieldAggregator;
  /**  Timestamp of this event  */
  timestamp: Scalars['BigInt']['output'];
  /**  Address that received tokens  */
  to: Scalars['String']['output'];
};
// ...
export type Position = {
  __typename?: 'Position';
  /**  Account associated with the position  */
  account: Account;
  claimedSummerToken: Scalars['BigInt']['output'];
  claimedSummerTokenNormalized: Scalars['BigDecimal']['output'];
  /**  Creation block number  */
  createdBlockNumber: Scalars['BigInt']['output'];
  /**  Creation timestamp  */
  createdTimestamp: Scalars['BigInt']['output'];
  dailySnapshots: Array<PositionDailySnapshot>;
  deposits: Array<Deposit>;
  hourlySnapshots: Array<PositionHourlySnapshot>;
  /**  Unique identifier for the position  */
  id: Scalars['ID']['output'];
  /**  Balance of the input token for the position  */
  inputTokenBalance: Scalars['BigInt']['output'];
  /**  Normalized supply of the input token  */
  inputTokenBalanceNormalized: Scalars['BigDecimal']['output'];
  /**  Normalized supply of the input token in USD  */
  inputTokenBalanceNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Sum of deposits in the position - in input token precision */
  inputTokenDeposits: Scalars['BigInt']['output'];
  /**  Sum of deposits in the position - in input token precision normalized */
  inputTokenDepositsNormalized: Scalars['BigDecimal']['output'];
  /**  Sum of deposits in the position - in input token precision normalized to USD */
  inputTokenDepositsNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Sum of withdrawals in the position - in input token precision */
  inputTokenWithdrawals: Scalars['BigInt']['output'];
  /**  Sum of withdrawals in the position - in input token precision normalized */
  inputTokenWithdrawalsNormalized: Scalars['BigDecimal']['output'];
  /**  Sum of withdrawals in the position - in input token precision normalized to USD */
  inputTokenWithdrawalsNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Supply of the output token for the position  */
  outputTokenBalance: Scalars['BigInt']['output'];
  stakedEvents: Array<Staked>;
  /**  Staked balance of the input token for the position  */
  stakedInputTokenBalance: Scalars['BigInt']['output'];
  /**  Normalized staked balance of the input token  */
  stakedInputTokenBalanceNormalized: Scalars['BigDecimal']['output'];
  /**  Normalized staked balance of the input token in USD  */
  stakedInputTokenBalanceNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Staked balance of the output token for the position  */
  stakedOutputTokenBalance: Scalars['BigInt']['output'];
  unstakedEvents: Array<Unstaked>;
  /**  Unstaked balance of the input token for the position  */
  unstakedInputTokenBalance: Scalars['BigInt']['output'];
  /**  Normalized unstaked balance of the input token  */
  unstakedInputTokenBalanceNormalized: Scalars['BigDecimal']['output'];
  /**  Normalized unstaked balance of the input token in USD  */
  unstakedInputTokenBalanceNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Unstaked balance of the output token for the position  */
  unstakedOutputTokenBalance: Scalars['BigInt']['output'];
  /**  Vault where the position is held  */
  vault: Vault;
  weeklySnapshots: Array<PositionWeeklySnapshot>;
  withdrawals: Array<Withdraw>;
};
// ...
export type PositionDailySnapshot = {
  __typename?: 'PositionDailySnapshot';
  /**  { Position ID }-{ # of days since Unix epoch time }  */
  id: Scalars['ID']['output'];
  /**  Amount of input token in the position  */
  inputTokenBalance: Scalars['BigInt']['output'];
  /**  Input token balance in the position normalized */
  inputTokenBalanceNormalized: Scalars['BigDecimal']['output'];
  /**  Amount of input token in the position normalized to USD  */
  inputTokenBalanceNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Sum of deposits in the position - in input token precision */
  inputTokenDeposits: Scalars['BigInt']['output'];
  /**  Sum of deposits in the position - in input token precision normalized */
  inputTokenDepositsNormalized: Scalars['BigDecimal']['output'];
  /**  Sum of deposits in the position - in input token precision normalized to USD */
  inputTokenDepositsNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Sum of withdrawals in the position - in input token precision */
  inputTokenWithdrawals: Scalars['BigInt']['output'];
  /**  Sum of withdrawals in the position - in input token precision normalized */
  inputTokenWithdrawalsNormalized: Scalars['BigDecimal']['output'];
  /**  Sum of withdrawals in the position - in input token precision normalized to USD */
  inputTokenWithdrawalsNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Amount of output token in the position  */
  outputTokenBalance: Scalars['BigInt']['output'];
  /**  The position this snapshot belongs to  */
  position: Position;
  /**  Timestamp of this snapshot  */
  timestamp: Scalars['BigInt']['output'];
};
// ...
export type PositionHourlySnapshot = {
  __typename?: 'PositionHourlySnapshot';
  /**  { Position ID }-{ # of hours since Unix epoch time }  */
  id: Scalars['ID']['output'];
  /**  Amount of input token in the position  */
  inputTokenBalance: Scalars['BigInt']['output'];
  /**  Input token balance in the position normalized */
  inputTokenBalanceNormalized: Scalars['BigDecimal']['output'];
  /**  Amount of input token in the position normalized to USD  */
  inputTokenBalanceNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Sum of deposits in the position - in input token precision */
  inputTokenDeposits: Scalars['BigInt']['output'];
  /**  Sum of deposits in the position - in input token precision normalized */
  inputTokenDepositsNormalized: Scalars['BigDecimal']['output'];
  /**  Sum of deposits in the position - in input token precision normalized to USD */
  inputTokenDepositsNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Sum of withdrawals in the position - in input token precision */
  inputTokenWithdrawals: Scalars['BigInt']['output'];
  /**  Sum of withdrawals in the position - in input token precision normalized */
  inputTokenWithdrawalsNormalized: Scalars['BigDecimal']['output'];
  /**  Sum of withdrawals in the position - in input token precision normalized to USD */
  inputTokenWithdrawalsNormalizedInUSD: Scalars['BigDecimal']['output'];
  /**  Amount of output token in the position  */
  outputTokenBalance: Scalars['BigInt']['output'];
  /**  The position this snapshot belongs to  */
  position: Position;
  /**  Timestamp of this snapshot  */
  timestamp: Scalars['BigInt']['output'];
};
// ...
export type Staked = Event & {
  __typename?: 'Staked';
  /**  Amount of token staked in native units  */
  amount: Scalars['BigInt']['output'];
  /**  Amount of token staked in USD  */
  amountUSD: Scalars['BigDecimal']['output'];
  /**  Token staked  */
  asset: Token;
  /**  Block number of this event  */
  blockNumber: Scalars['BigInt']['output'];
  /**  Address that sent tokens  */
  from: Scalars['String']['output'];
  /**  Transaction hash of the transaction that emitted this event  */
  hash: Scalars['String']['output'];
  /**  { Transaction hash }-{ Log index } */
  id: Scalars['ID']['output'];
  /**  Amount of input token in the position  */
  inputTokenBalance: Scalars['BigInt']['output'];
  /**  Amount of input token in the position in USD  */
  inputTokenBalanceNormalizedUSD: Scalars['BigDecimal']['output'];
  /**  Event log index. For transactions that don't emit event, create arbitrary index starting from 0  */
  logIndex: Scalars['Int']['output'];
  /**  Position that this stake belongs to  */
  position: Position;
  /**  The protocol this transaction belongs to  */
  protocol: YieldAggregator;
  /**  Timestamp of this event  */
  timestamp: Scalars['BigInt']['output'];
  /**  Address that received tokens  */
  to: Scalars['String']['output'];
  /**  The vault involving this transaction  */
  vault: Vault;
};
// ...
export type Unstaked = Event & {
  __typename?: 'Unstaked';
  /**  Amount of token unstaked in native units  */
  amount: Scalars['BigInt']['output'];
  /**  Amount of token unstaked in USD  */
  amountUSD: Scalars['BigDecimal']['output'];
  /**  Token withdrawn  */
  asset: Token;
  /**  Block number of this event  */
  blockNumber: Scalars['BigInt']['output'];
  /**  Address that sent tokens  */
  from: Scalars['String']['output'];
  /**  Transaction hash of the transaction that emitted this event  */
  hash: Scalars['String']['output'];
  /**  { Transaction hash }-{ Log index } */
  id: Scalars['ID']['output'];
  /**  Amount of input token in the position  */
  inputTokenBalance: Scalars['BigInt']['output'];
  /**  Amount of input token in the position in USD  */
  inputTokenBalanceNormalizedUSD: Scalars['BigDecimal']['output'];
  /**  Event log index. For transactions that don't emit event, create arbitrary index starting from 0  */
  logIndex: Scalars['Int']['output'];
  /**  Position that this unstake belongs to  */
  position: Position;
  /**  The protocol this transaction belongs to  */
  protocol: YieldAggregator;
  /**  Timestamp of this event  */
  timestamp: Scalars['BigInt']['output'];
  /**  Address that received tokens  */
  to: Scalars['String']['output'];
  /**  The vault involving this transaction  */
  vault: Vault;
};